requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
branca==0.8.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chart-studio==1.1.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlover==0.3.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cufflinks==0.17.3
cycler==0.12.1
databricks-sdk==0.44.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
durationpy==0.9
executing==2.1.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
folium==0.19.0
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2025.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.38.0
googleapis-common-protos==1.69.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==32.0.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.19
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.20.2
mlflow-skinny==2.20.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.15.2
nest-asyncio==1.6.0
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.69.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.1
parso==0.8.4
patsy==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
platformdirs==4.3.6
plotly==5.22.0
posthog==3.23.0
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==5.29.4
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
setuptools==78.1.0
shap==0.46.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
streamlit==1.34.0
streamlit_folium==0.23.2
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
xgboost==2.1.4
xyzservices==2024.9.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
langchain_openai
